+++ CHANGELOG.md
@@ -4,6 +4,14 @@ yii2-enumerables change Log
 dev-master
 ----------
 
+version 1.1.0
+-------------
+**Date:** 05-Aug-2015
+
+- Enh: Added class for generates a list of months.
+- Enh: Added class for generates a list of days.
+
+
 version 1.0.0
 -------------
 **Date:** 18-Jul-2015